+++ packages/terra-button/tests/jest/Button.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 
-import Button, { IconTypes } from '../../src/Button';
+import Button from '../../src/Button';
 
 // Snapshot Tests
 it('should render a neutral button', () => {
@@ -51,7 +51,7 @@ it('should render a button with type equal to submit', () => {
 
 it('should render an informative icon button', () => {
   const testElement = <img alt="icon" />;
-  const button = shallow(<Button iconType={IconTypes.INFORMATIVE} icon={testElement} isIconOnly text="isIconOnly" />);
+  const button = shallow(<Button icon={testElement} isIconOnly text="isIconOnly" />);
   expect(button).toMatchSnapshot();
 });